SIA-497 moved processor file
